+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 CHANGELOG
 =========
 
-0.1.7 (--.--.2013)
+0.1.7 (08.10.2013)
 ------------------
 * Added support PDO options and attributes
 * Added Redis adapter to Cache package
 * Added JsonSerializable interface to Db\Row
 * Added JsonSerializable interface to View\View
 * Added support HTTP PATCH method to HTTPRequest
+* Added Controller package with Rest and Crud controllers
+* Refactoring Crud package
 * Refactoring render() method of Application
 
 0.1.6 (13.08.2013)